Extract BaseSelectListItem component
