+++ frontend/src/metabase/components/select-list/SelectListItem.jsx
@@ -3,39 +3,24 @@ import PropTypes from "prop-types";
 import _ from "underscore";
 
 import { iconPropTypes } from "metabase/components/Icon";
-import { useScrollOnMount } from "metabase/hooks/use-scroll-on-mount";
 
-import { ItemRoot, ItemIcon, ItemTitle } from "./SelectListItem.styled";
+import { BaseSelectListItem } from "./BaseSelectListItem";
+import { ItemIcon, ItemTitle } from "./SelectListItem.styled";
+
+const iconPropType = PropTypes.oneOfType([
+  PropTypes.string,
+  PropTypes.shape(iconPropTypes),
+]);
 
 const propTypes = {
-  id: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  icon: PropTypes.oneOfType([PropTypes.string, PropTypes.shape(iconPropTypes)])
-    .isRequired,
+  ...BaseSelectListItem.propTypes,
+  icon: iconPropType.isRequired,
   iconColor: PropTypes.string,
-  onSelect: PropTypes.func.isRequired,
-  isSelected: PropTypes.bool,
-  rightIcon: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.shape({
-      name: PropTypes.name,
-    }),
-  ]),
-  size: PropTypes.oneOf(["small", "medium"]),
-  className: PropTypes.string,
+  rightIcon: iconPropType,
 };
 
-export function SelectListItem({
-  id,
-  name,
-  icon,
-  onSelect,
-  isSelected = false,
-  rightIcon,
-  size = "medium",
-  className,
-}) {
-  const ref = useScrollOnMount();
+export function SelectListItem(props) {
+  const { name, icon, rightIcon } = props;
 
   const iconProps = _.isObject(icon) ? icon : { name: icon };
   const rightIconProps = _.isObject(rightIcon)
@@ -43,20 +28,11 @@ export function SelectListItem({
     : { name: rightIcon };
 
   return (
-    <ItemRoot
-      innerRef={isSelected ? ref : null}
-      isSelected={isSelected}
-      role="menuitem"
-      tabIndex={0}
-      size={size}
-      onClick={() => onSelect(id)}
-      onKeyDown={e => e.key === "Enter" && onSelect(id)}
-      className={className}
-    >
+    <BaseSelectListItem {...props}>
       <ItemIcon color="brand" {...iconProps} />
       <ItemTitle>{name}</ItemTitle>
       {rightIconProps.name && <ItemIcon {...rightIconProps} />}
-    </ItemRoot>
+    </BaseSelectListItem>
   );
 }
