+++ frontend/src/metabase/components/select-list/SelectListItem.styled.jsx
@@ -28,10 +28,7 @@ const VERTICAL_PADDING_BY_SIZE = {
   medium: "0.75rem",
 };
 
-export const ItemRoot = styled.li`
-  display: grid;
-  grid-template-columns: min-content 1fr min-content;
-  gap: 0.5rem;
+export const BaseItemRoot = styled.li`
   align-items: center;
   cursor: pointer;
   padding: ${props => VERTICAL_PADDING_BY_SIZE[props.size]} 0.5rem;
@@ -47,4 +44,10 @@ export const ItemRoot = styled.li`
   &:hover {
     ${activeItemCss}
   }
+`;
+
+export const ItemRoot = styled(BaseItemRoot)`
+  display: grid;
+  grid-template-columns: min-content 1fr min-content;
+  gap: 0.5rem;
 `;
